# [REVIEW] Raise temporary error for `decimal128` types in parquet reader