--- a/python/cudf/cudf/io/parquet.py
+++ b/python/cudf/cudf/io/parquet.py
@@ -422,10 +422,12 @@ def read_parquet(
             else:
                 arrow_types = metadata.schema.to_arrow_schema().types
                 for arrow_type in arrow_types:
-                    if isinstance(arrow_type, (pa.ListType, pa.StructType)):
+                    if isinstance(arrow_type, pa.ListType):
                         val_field_types = arrow_type.value_field.flatten()
                         for val_field_type in val_field_types:
                             _check_decimal128_type(val_field_type.type)
+                    elif isinstance(arrow_type, pa.StructType):
+                        _ = cudf.StructDtype.from_arrow(arrow_type)
                     else:
                         _check_decimal128_type(arrow_type)