@galipremsagar 2021-12-07
> Do we know anything about the performance implications of this? I don't know how expensive functions like `to_arrow_schema` and `pq.read_metadata` are, but it does seem like significant error-checking work to be done up front. Would it be possible to try reading the file, catching whatever error we get back from that, and then checking for 128-bit columns after the fact?

Here is a benchmark of reading 1000 columns dataframe:

```ipython
In [3]: df = cudf.DataFrame(columns=range(1, 1000))

In [4]: df
Out[4]: d
Empty DataFrame
Columns: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 980, 981, 982, 983, 984, 985, 986, 987, 988, 989, 990, 991, 992, 993, 994, 995, 996, 997, 998, 999]
Index: []

In [5]: df[1] = ['a', 'b', 'c']

In [6]: df
Out[6]: 
  1     2     3     4     5     6     7     8     9     10    11    12    13    14    15   ...   985   986   987   988   989   990   991   992   993   994   995   996   997   998   999
0   a  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  ...  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>
1   b  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  ...  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>
2   c  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  ...  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>

[3 rows x 999 columns]

In [10]: df.columns = df.columns.astype('str')

In [11]: df.to_parquet('long.parquet')

# Before
In [12]: %timeit cudf.read_parquet("long.parquet")
151 ms ± 1.61 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# This PR
In [3]: %timeit cudf.read_parquet("long.parquet")
171 ms ± 4.53 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```