@galipremsagar 2021-12-07
>  I wouldn't want to release 22.02 with this performance regression in it though. Do you or @vuule know what the timeline is?

Yup, it is being targeted for 22.02