@vuule 2021-12-07
> A 10-15% performance regression is pretty undesirable, but not as big a deal if it's temporary

IMO it's fine, as this is 15% with an empty dataframe, and the overhead does not increase with the number of rows. So it should be negligible in most cases.